requirements.txt
motor
python-dotenv
aiogram